requirements.txt
SQLAlchemy==2.0.28
psycopg2==2.9.9
python-dotenv==1.0.1
attrs==23.2.0

